+++ HopperRender/HopperRenderSettings.h
@@ -24,9 +24,11 @@ private:
 	int m_iNumIterations; // Number of iterations to use in the optical flow calculation (0: As many as possible)
 	int m_iFrameBlurKernelSize; // The size of the blur kernel used to blur the source frames before calculating the optical flow
 	int m_iFlowBlurKernelSize; // The size of the blur kernel used to blur the offset calculated by the optical flow
+	int m_iSceneChangeThreshold; // The threshold used to determine whether a scene change has occurred
 	int m_iIntActiveState; // The state of the filter (0: Deactivated, 1: Not Needed, 2: Active, 3: Too Slow)
 	double m_dSourceFPS; // The source frames per second
 	int m_iNumSteps; // Number of steps executed to find the ideal offset (limits the maximum offset distance per iteration)
+	int m_iCurrentSceneChange; // How many pixel differences are currently detected
 	bool m_bIsInitialized; // Used to ignore startup messages
 	unsigned char m_cRefreshCounter; // Counts the number of times the settings page has been refreshed
 	SettingsInterface* m_pSettingsInterface; // The custom interface on the filter